Added 400 bad request on properties<id>

+++ flaskapp/routes/properties.py
@@ -2,11 +2,9 @@ import flask
 from flask import request, jsonify
 from flask_api import status
 import pymysql
-
 from DBGateway import DBGateway
 
 currentDir = ""
-
 app = flask.current_app
 bp = flask.Blueprint("properties", __name__)
 
@@ -39,47 +37,62 @@ def add_property():
             cur.close()
             conn.close()
         #end try/except/finally
+
     elif request.method == 'POST':
-        req_data = request.get_json(force=True)
-        address = req_data['address']
-        city = req_data['city']
-        state = req_data['state']
-        zip_code = req_data['zip']
+        headers = request.headers
+        auth = headers.get("X-Api-Key")
+        if auth == 'cs4783FTW':
+            req_data = request.get_json(force=True)
+            address = req_data['address']
+            city = req_data['city']
+            state = req_data['state']
+            zip_code = req_data['zip']
 
-        if 1 <= len(address) <= 200 and len(state) == 2 and 1 <= len(city) <= 50 and 5 <= len(zip_code) <= 10:
-            cur = None
-            conn = None
-            try:
-                sql = "INSERT INTO tbl_property(ID, address, city, state, zip) VALUES(NULL, %s, %s, %s, %s)"
-                data = (address, city, state, zip_code)
-                gateway = DBGateway()
-                conn = gateway.get_connection()
-                cur = conn.cursor()
-                cur.execute(sql, data)
-                conn.commit()
-                return prepare_response("message", "added"), status.HTTP_200_OK
-            except Exception as e:
-                print(e)
-            finally:
-                cur.close()
-                conn.close()
-            #end try/except/finally
-        #end if
+            if 1 <= len(address) <= 200 and len(state) == 2 and 1 <= len(city) <= 50 and 5 <= len(zip_code) <= 10:
+                cur = None
+                conn = None
+                try:
+                    sql = "INSERT INTO tbl_property(ID, address, city, state, zip) VALUES(NULL, %s, %s, %s, %s)"
+                    data = (address, city, state, zip_code)
+                    gateway = DBGateway()
+                    conn = gateway.get_connection()
+                    cur = conn.cursor()
+                    cur.execute(sql, data)
+                    conn.commit()
+                    return prepare_response("message", "added"), status.HTTP_200_OK
+                except Exception as e:
+                    print(e)
+                finally:
+                    cur.close()
+                    conn.close()
+                #end try/except/finally
+            #end if
 
-        if len(address) < 1 or len(address) > 200:
-            return prepare_response("message", "address is not between 1 and 200 characters"), status.HTTP_400_BAD_REQUEST
-        if len(city) < 1 or len(city) > 50:
-            return prepare_response("message", "city is not between 1 and 50 characters"), status.HTTP_400_BAD_REQUEST
-        if len(state) != 2:
-            return prepare_response("message", "state is not 2 characters"), status.HTTP_400_BAD_REQUEST
-        if len(zip_code) < 5 or len(zip_code) > 10:
-            return prepare_response("message", "zip is not between 5 and 10 characters"), status.HTTP_400_BAD_REQUEST
-    #end if/elif
+            if len(address) < 1 or len(address) > 200:
+                return prepare_response("message", "address is not between 1 and 200 characters"), status.HTTP_400_BAD_REQUEST
+            if len(city) < 1 or len(city) > 50:
+                return prepare_response("message", "city is not between 1 and 50 characters"), status.HTTP_400_BAD_REQUEST
+            if len(state) != 2:
+                return prepare_response("message", "state is not 2 characters"), status.HTTP_400_BAD_REQUEST
+            if len(zip_code) < 5 or len(zip_code) > 10:
+                return prepare_response("message", "zip is not between 5 and 10 characters"), status.HTTP_400_BAD_REQUEST
+            #end if
+        else:
+            return jsonify({"message": "ERROR: Unauthorized"}), 401
+        #end if/else
 #end properties
 
 
-@bp.route('/properties/<int:property_id>', methods=['GET', 'DELETE', 'PUT'])
+@bp.route('/properties/<property_id>', methods=['GET', 'DELETE', 'PUT'])
 def id_property(property_id):
+    try:
+        property_id = int(property_id)
+        if (not isinstance(property_id, int)):
+            return prepare_response("message", "ID is not an integer"), status.HTTP_400_BAD_REQUEST
+    except ValueError:
+        return prepare_response("message", "ID is not an integer"), status.HTTP_400_BAD_REQUEST
+    #end try/catch
+    
     if request.method == 'GET':
         cur = None
         conn = None
@@ -101,6 +114,7 @@ def id_property(property_id):
             cur.close()
             conn.close()
         # end try/except/finally
+
     elif request.method == 'PUT':
         headers = request.headers
         auth = headers.get("X-Api-Key")
@@ -151,28 +165,35 @@ def id_property(property_id):
         else:
             return jsonify({"message": "ERROR: Unauthorized"}), 401
         #end if/else
+
     elif request.method == 'DELETE':
-        cur = None
-        conn = None
-        try:
-            gateway = DBGateway()
-            conn = gateway.get_connection()
-            cur = conn.cursor(pymysql.cursors.DictCursor)
-            # check to see if the property exists
-            cur.execute("SELECT * FROM tbl_property WHERE ID=%s", property_id)
-            rows = cur.fetchall()
-            if not rows:
-                return prepare_response("message", "ID does not exist in database."), status.HTTP_404_NOT_FOUND
-            # delete the property if it does
-            cur.execute("DELETE FROM tbl_property WHERE ID=%s", property_id)
-            conn.commit()
-            return prepare_response("message", "deleted"), status.HTTP_200_OK
-        except Exception as e:
-            print(e)
-        finally:
-            cur.close()
-            conn.close()
-        # end try/except/finally
+        headers = request.headers
+        auth = headers.get("X-Api-Key")
+        if auth == 'cs4783FTW':
+            cur = None
+            conn = None
+            try:
+                gateway = DBGateway()
+                conn = gateway.get_connection()
+                cur = conn.cursor(pymysql.cursors.DictCursor)
+                # check to see if the property exists
+                cur.execute("SELECT * FROM tbl_property WHERE ID=%s", property_id)
+                rows = cur.fetchall()
+                if not rows:
+                    return prepare_response("message", "ID does not exist in database."), status.HTTP_404_NOT_FOUND
+                # delete the property if it does
+                cur.execute("DELETE FROM tbl_property WHERE ID=%s", property_id)
+                conn.commit()
+                return prepare_response("message", "deleted"), status.HTTP_200_OK
+            except Exception as e:
+                print(e)
+            finally:
+                cur.close()
+                conn.close()
+            # end try/except/finally
+        else:
+            return jsonify({"message": "ERROR: Unauthorized"}), 401
+        #end if/else
     # end if/elif
 #end properties property_id
